// apps/sage-sync/src/lib/sage/types.ts
type SageRecord = {
  RECORD_MODIFY_DATE: string;
  RECORD_CREATE_DATE: string;
  RECORD_DELETED: 0 | 1;
};
// ...
export type STOCK_TRAN = SageRecord & {
  TRAN_NUMBER: number;
  STOCK_CODE: string;
  TYPE: string;
  DATE: string;
  REFERENCE: string;
  REFERENCE_NUMERIC: number;
  DETAILS: string;
  QUANTITY: number;
  COST_PRICE: number;
  SALES_PRICE: number;
  QTY_USED: number;
  ISP_REFERENCE: number;
  ARTEFACT_TYPE: number;
  GDN_NUMBER: number;
  GRN_NUMBER: number;
};

export type GDN_ITEM = SageRecord & {
  UNIQUE_ID: number;
  GDN_NUMBER: number;
  ITEM_NUMBER: number;
  ORDER_NUMBER: number;
  ORDER_ITEM: number;
  ACCOUNT_REF: string;
  CUSTOMER_GDN_NUMBER: string;
  STOCK_CODE: string;
  DESCRIPTION: string;
  DATE: string;
  QTY_DESPATCHED: number;
  QTY_ON_ORDER: number;
  PRINTED_FLAG: number;
};

export type GRN_ITEM = SageRecord & {
  GRN_NUMBER: number;
  ITEM_NUMBER: number;
  ORDER_NUMBER: number;
  ORDER_ITEM: number;
  ACCOUNT_REF: string;
  SUPPLIER_GRN_NUMBER: string;
  STOCK_CODE: string;
  DESCRIPTION: string;
  DATE: string;
  QTY_RECEIVED: number;
  QTY_ON_ORDER: number;
};